* Small shadow effect for inworld scene dialog widgets * Fade in on show event * Should not drop FPS as the animations did. If people have problems in debug we can put these only for debug. Try them out!
git-svn-id: https://realxtend-naali.googlecode.com/svn/trunk@2241 5b2332b8-efa3-11de-8684-7d64432d61a3

+++ UiModule/UiProxyWidget.h
@@ -37,7 +37,7 @@ namespace UiServices
         QGraphicsItemAnimation *show_animation_;
 
     private slots:
-        void AnimationStep(int step);
+        void AnimationStep(qreal step);
 
     signals:
         void Visible(bool);
